requirements.txt
asgiref==3.8.1
asttokens==3.0.0
black==25.1.0
boto3==1.20.26
botocore==1.23.54
certifi==2023.11.17
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.8
coverage==7.6.12
cryptography==41.0.7
decorator==5.2.1
diff-match-patch==20241021
distlib==0.3.9
dj-database-url==2.1.0
Django==4.2.19
django-anymail==10.2
django-cors-headers==3.14.0
django-import-export==4.3.7
django-jazzmin==2.6.0
django-storages==1.12.3
django-tinymce==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
environs==10.0.0
executing==2.2.0
filelock==3.17.0
flake8==7.1.2
gunicorn==21.2.0
h11==0.14.0
identify==2.6.8
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.32.0
isort==6.0.0
jedi==0.19.2
jmespath==0.10.0
marshmallow==3.20.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.1.0
prompt_toolkit==3.0.50
psycopg2==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.12.1
pycparser==2.21
pyflakes==3.2.0
Pygments==2.19.1
PyJWT==2.6.0
pytest==8.3.5
pytest-cov==6.0.0
pytest-django==4.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2025.1
PyYAML==6.0.2
requests==2.31.0
ruff==0.9.8
s3transfer==0.5.2
setuptools==75.8.0
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.3
stack-data==0.6.3
stripe==7.9.0
tablib==3.8.0
traitlets==5.14.3
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.34.0
virtualenv==20.29.2
wcwidth==0.2.13
wheel==0.45.1
